problems/src/test/java/_1001_1050/_1026_Maximum_Difference_Between_Node_and_Ancestor/SolutionITest.java
package _1001_1050._1026_Maximum_Difference_Between_Node_and_Ancestor;

import definition.binaryTree.TreeNode;
import definition.binaryTree.TreeNodes;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.*;

/**
 * @author jacka
 * @version 1.0 on 10/19/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final TreeNode input = TreeNodes.getTreeLevelOrder(8, 3, 10, 1, 6, null, 14, null, null, 4, 7, 13);
        final int expected = 7, actual = solution.maxAncestorDiff(input);
    assertEquals(expected, actual);
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}